# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoRecicla Bogotá API - Recycling Management System for Residential Complexes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: complexId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # RESIDENTIAL COMPLEXES
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, documentNumber, email, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        documentNumber:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: string
        apartment:
          type: ["string", "null"]
        complexId:
          type: ["integer", "null"]
        complexName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, resident]
        status:
          type: string
          enum: [active, inactive]
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [fullName, documentNumber, email, role, status, password]
      properties:
        fullName:
          type: string
        documentNumber:
          type: string
        phone:
          type: string
        email:
          type: string
        apartment:
          type: string
        complexId:
          type: integer
        role:
          type: string
          enum: [admin, resident]
        status:
          type: string
          enum: [active, inactive]
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        documentNumber:
          type: string
        phone:
          type: string
        email:
          type: string
        apartment:
          type: string
        complexId:
          type: ["integer", "null"]
        role:
          type: string
          enum: [admin, resident]
        status:
          type: string
          enum: [active, inactive]
    UserListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # COMPLEX
...